+++ app/controllers/api/ui/v1/course_events_controller.rb
@@ -17,7 +17,7 @@ class Api::Ui::V1::CourseEventsController < Api::Ui::BaseController
       if params[:state].downcase == 'active'
         @course_events = @course_events
                     .where(state: :active)
-                    .where('event_date > ?', Time.zone.now+EVENT_TIME_OFFSET)
+                    # .where('event_date > ?', Time.zone.now+EVENT_TIME_OFFSET)
       elsif params[:state].downcase == 'closed'
         @course_events = @course_events
                              .where(state: :closed)