+++ app/javascript/users/mixins/homepage_content_mixin.js
@@ -10,7 +10,11 @@ export const homepageContentMixin = {
       return this.current_organization.homepage
     },
     organizationEmail() {
-      return this.current_organization.email || 'support@insiemento.com'
+      return this.current_organization.email || 'info@insiemento.com'
+    },
+    organizationLogo() {
+      // if (this.current_organization.logo) console.log('ATTENZIONE: caricare il logo per la compagnia ', this.current_organization.id)
+      return this.current_organization.logo || require('../../../assets/images/logo_200.png') //
     },
   },
 }