+++ app/models/course.rb
@@ -8,6 +8,14 @@ class Course < ApplicationRecord
   has_many :trainers, through: :course_schedules
   has_many :rooms, through: :course_schedules
 
+  before_validation :set_default
+
   STATES = { just_made: 10, active: 20, suspended: 30}
   enum state: STATES
+
+  private
+
+  def set_default
+    self.state ||= :active
+  end
 end