+++ app/models/course_event.rb
@@ -13,10 +13,18 @@ class CourseEvent < ApplicationRecord
   validates :course_id, uniqueness: { scope: [:room_id, :trainer_id, :event_date],
                                  message: 'already exists' }
 
+  before_validation :set_default
+
   STATES = { just_made: 10, active: 20, suspended: 30, closed: 40}
   enum state: STATES
 
   def attendees_count
     attendees.count
+  end
+
+  private
+
+  def set_default
+    self.state ||= :active
   end
 end