+++ app/models/room.rb
@@ -5,7 +5,14 @@ class Room < ApplicationRecord
   has_many :course_schedules
   has_many :courses, through: :course_schedules
 
+  before_validation :set_default
+
   STATES = { just_made: 10, active: 20, suspended: 30}
   enum state: STATES
 
+  private
+
+  def set_default
+    self.state ||= :active
+  end
 end