+++ spec/factories/rooms.rb
@@ -1,0 +1,7 @@
+FactoryBot.define do
+  factory :room do
+    organization
+    sequence(:name) { |n| "Room#{n}" }
+    max_attendees { 24 }
+  end
+end